% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mnist_test}
\alias{mnist_test}
\title{MNIST dataset - testing}
\format{List with 3 elements:
\enumerate{
\item \strong{n}: number of elements
\item \strong{x}: n x 784 matrix. Each row corresponds to an observation and each
column corresponds to a pixel in the image.
\item \strong{y}: n-entry vector. Responds to number between 0 and 9, which is the
digit written down. Preprocessed into an R factor.
}}
\source{
\href{http://yann.lecun.com/exdb/mnist/}{Yann Lecun}
}
\usage{
mnist_test
}
\description{
60000 observations of 28x28 image data. Corresponds to digits between
0 and 9.
}
\keyword{datasets}
