% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnist.R
\name{open_binary_file}
\alias{open_binary_file}
\title{Open Gzipped Binary File at URL}
\usage{
open_binary_file(filename, base_url = mnist_url, verbose = FALSE)
}
\arguments{
\item{filename}{Name of the file to open.}

\item{base_url}{URL of the resource containing the file.}

\item{verbose}{If \code{TRUE}, generate a diagnostic message when the file
is opened.}
}
\value{
Opened connection to the file.
}
\description{
Opens a file at a specified URL and returns the connection for further
processing. Callers must close the connection when they're done.
}
