% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnist.R
\name{parse_files}
\alias{parse_files}
\title{Parse Image and Label File Pair}
\usage{
parse_files(image_filename, label_filename, base_url = mnist_url,
  verbose = FALSE)
}
\arguments{
\item{image_filename}{The image filename.}

\item{label_filename}{The label filename corresponding to the images in
\code{image_filename}.}

\item{base_url}{URL of the resource containing the files.}

\item{verbose}{If \code{TRUE}, generate a diagnostic message as files are
downloaded.}
}
\value{
Data frame containing images and labels.
}
\description{
Downloads an image file and a corresponding label file, combining
them into a data frame.
}

