% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnist.R
\name{parse_image_file}
\alias{parse_image_file}
\title{Parse Image File}
\usage{
parse_image_file(filename, base_url = mnist_url, verbose = verbose)
}
\arguments{
\item{filename}{The image filename.}

\item{base_url}{URL of the resource containing the \code{filename}.}

\item{verbose}{If \code{TRUE}, generate a diagnostic message as files are
downloaded.}
}
\value{
Vector of integers representing the digits.
}
\description{
Downloads a gzipped MNIST image file.
}

