% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnist.R
\name{parse_label_file}
\alias{parse_label_file}
\title{Parse Label File}
\usage{
parse_label_file(filename, base_url = mnist_url, verbose = FALSE)
}
\arguments{
\item{filename}{The label filename.}

\item{base_url}{URL of the resource containing \code{filename}.}

\item{verbose}{If \code{TRUE}, generate a diagnostic message as files are
downloaded.}
}
\value{
Vector of integers representing the digits.
}
\description{
Downloads a gzipped MNIST label file.
}
