% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mnist.R
\name{parse_mnist}
\alias{parse_mnist}
\title{Download the MNIST database of handwritten digits.}
\format{A data frame with 785 variables:

\describe{
\item{\code{px1}, \code{px2}, \code{px3} ... \code{px784}}{Integer pixel value, from 0 (white) to 255 (black).}
\item{\code{Label}}{The digit represented by the image, in the range 0-9.}
}

Pixels are organized row-wise. The \code{Label} variable is stored as a
factor.

There are 70,000 digits in the data set. The first 60,000 are the training
set, as found in the \code{train-images-idx3-ubyte.gz} file. The remaining
10,000 are the test set, from the \code{t10k-images-idx3-ubyte.gz} file.

For more information see \url{http://yann.lecun.com/exdb/mnist}.}
\usage{
parse_mnist()
}
\description{
Downloads the image and label files for the training and test datasets from
\url{http://yann.lecun.com/exdb/mnist} and converts them to a data frame.
}
\note{
Originally based on a function by Brendan O'Connor.
}
\examples{
# download the files and parse them
mnist <- parse_mnist()

# first 60,000 instances are the training set
mnist_train <- head(mnist, 60000)

# the remaining 10 000 are the test set
mnist_test <- tail(mnist, 10000)

# PCA on 1000 random training examples
mnist_r1000 <- mnist_train[sample(nrow(mnist_train), 1000), ]
pca <- princomp(mnist_r1000[,1:784], scores = TRUE)

# plot the scores of the first two components
plot(pca$scores[,1:2], t='n')
text(pca$scores[,1:2], labels = mnist_r1000$Label,
     col = rainbow(length(levels(mnist_r1000$Label)))[mnist_r1000$Label])
}

