% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_digit.R
\name{show_digit}
\alias{show_digit}
\title{Plots an MNIST digit}
\usage{
show_digit(arr784, col = gray(12:1/12), ...)
}
\arguments{
\item{arr784}{an 'image'}

\item{col}{what color to draw the image iwth}

\item{...}{other options passed to \code{image}}
}
\description{
Works by default for the MNIST objects as established here.
}
\examples{
show_digit(MNIST::mnist_train$x[30,])
}
