% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mnist.R
\name{show_digit}
\alias{show_digit}
\title{Visualize a digit from the MNIST dataset.}
\usage{
show_digit(df, n, col = gray(12:1/12), ...)
}
\arguments{
\item{df}{Data frame containing MNIST digits.}

\item{n}{row index of the digit to display.}

\item{col}{list of colors to use in the display.}

\item{...}{Other arguments passed onto the \code{image} function.}
}
\description{
Displays as an image the hand-written digit represented by the nth row in
a data frame.
}
\note{
Originally based on a function by Brendan O'Connor.
}
\examples{
# show the fifth digit
mnist <- download_mnist()
show_digit(mnist, 5)
}

