% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnist.R
\name{show_digit}
\alias{show_digit}
\title{Visualize MNIST digit.}
\usage{
show_digit(df, n, col = gray(12:1/12), ...)
}
\arguments{
\item{df}{Data frame containing MNIST digits.}

\item{n}{Row index of the digit to display.}

\item{col}{List of colors to use in the display.}

\item{...}{Other arguments passed onto the \code{\link[graphics]{image}}
function.}
}
\description{
Display an MNIST digit as an image the hand-written digit represented by the
nth row in a data frame.
}
\note{
Originally based on a function by Brendan O'Connor.
}
\examples{
\dontrun{
# show the fifth digit
mnist <- download_mnist()s
show_digit(mnist, 5)
}
}

